More correct names and packages for completion providers for Cargo.toml
