+++ toml/src/main/kotlin/org/rust/toml/completion/LocalCargoTomlDependencyCompletionProvider.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.toml.crates.local.completion
+package org.rust.toml.completion
 
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.completion.CompletionUtil
@@ -15,13 +15,11 @@ import com.intellij.codeInsight.lookup.LookupElementRenderer
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.components.service
 import org.rust.toml.StringValueInsertionHandler
-import org.rust.toml.completion.CargoNormalizedNamesPrefixMatcher
-import org.rust.toml.completion.TomlKeyValueCompletionProviderBase
 import org.rust.toml.crates.local.CratesLocalIndexService
 import org.rust.toml.crates.local.lastVersion
 import org.toml.lang.psi.TomlKeyValue
 
-class DependencyCratesLocalCompletionProvider : TomlKeyValueCompletionProviderBase() {
+class LocalCargoTomlDependencyCompletionProvider : TomlKeyValueCompletionProviderBase() {
     override fun completeKey(keyValue: TomlKeyValue, result: CompletionResultSet) {
         val prefix = CompletionUtil.getOriginalElement(keyValue.key)?.text ?: return
